Moved LastActivity to it's own class for readability.
Added a first rough BotDetector

Some final touches and moving Warning to it's own class

Fixed some small mistakes in ticks <--> seconds conversion

Fixed a stupid mistake that kept the interal TPS from updating and added
support for the immunePlayer set to BotDetector

Replace int with long

Removed a leftover that was not used

Fixed pom as needed for civcraft

Added a config option to enable/disable long bans, readded the
chatlistener, which only works at high TPS now and changed the look in
BotDetector to not use an additional iterator

Completly reworked how commands work and added commands to see a list of
banned players, suspected players and getter and setter for the
acceptabletps, the criticaltpschange and whether long bans are enabled
or not

Banned players are now logged to a file and read from that file on
startup so server crashes dont keep players banned longer than intended

Changed name of the command classes to start with an uppercase letter to
make Dan and the java convention happy. Also fixed some whitespace stuff

Couple of quick fixes to plugin.yml to pull version and plugin name from the pom, instead of duplicating places to change.

Console commands now work